Add support for jquery-pjax
Add `nprogress-pjax.js`, which triggers an `NProgress.start()` on the
`pjax:send` event, as well as an `NProgress.done()` on the
`pjax:complete` event. These two events are only fired when an actual
AJAX request is made, not when a cached page is loaded by the browser.
No event to call `NProgress.remove()` should be necessary.

Signed-off-by: David Celis <me@davidcel.is>